size_t removed due to compilator problems

+++ src/data_structures/headers/DynamicMedian.hpp
@@ -13,7 +13,7 @@ namespace ds
 
 		double GetMedian() const;
 		void Push(double number);
-		size_t GetSize() const;
+		int GetSize() const;
 
 	private:
 		void Balance();
